# packages/api-client/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "MonthBroadcastListDto": {
        "items": {
          "properties": {
            "id": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            },
            "school": {
              "type": "string"
            },
            "slug": {
              "type": "string"
            },
            "startsAt": {
              "format": "date-time",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
              "type": "string"
            },
            "state": {
              "enum": ["published", "live", "ended"],
              "type": "string"
            },
            "title": {
              "type": "string"
            }
          },
          "required": ["id", "slug", "title", "school", "startsAt", "state"],
          "type": "object"
        },
        "type": "array"
      },
...
      "PublicEventListingPageDto": {
        "properties": {
          "counts": {
            "properties": {
              "past": {
                "maximum": 9007199254740991,
                "minimum": 0,
                "type": "integer"
              },
              "upcoming": {
                "maximum": 9007199254740991,
                "minimum": 0,
                "type": "integer"
              }
            },
            "required": ["upcoming", "past"],
            "type": "object"
          },
          "data": {
            "items": {
              "anyOf": [
                {
                  "properties": {
                    "id": {
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                      "type": "string"
                    },
                    "school": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "speakers": {
                      "items": {
                        "properties": {
                          "name": {
                            "type": "string"
                          }
                        },
                        "required": ["name"],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "specialties": {
                      "items": {
                        "type": "string"
                      },
                      "type": "array"
                    },
                    "startsAt": {
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
                      "type": "string"
                    },
                    "state": {
                      "enum": ["published", "live"],
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "slug",
                    "title",
                    "school",
                    "startsAt",
                    "specialties",
                    "speakers",
                    "state"
                  ],
                  "type": "object"
                },
                {
                  "additionalProperties": false,
                  "properties": {
                    "id": {
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                      "type": "string"
                    },
                    "recording": {
                      "properties": {
                        "expectedBy": {
                          "anyOf": [
                            {
                              "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "posterUrl": {
                          "anyOf": [
                            {
                              "maxLength": 500,
                              "minLength": 1,
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "primaryKind": {
                          "anyOf": [
                            {
                              "enum": ["edited", "raw"],
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "secondaryKind": {
                          "anyOf": [
                            {
                              "enum": ["edited", "raw"],
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "state": {
                          "enum": ["montage", "raw-only", "preparing"],
                          "type": "string"
                        }
                      },
                      "required": [
                        "state",
                        "primaryKind",
                        "secondaryKind",
                        "posterUrl",
                        "expectedBy"
                      ],
                      "type": "object"
                    },
                    "school": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "speakers": {
                      "items": {
                        "properties": {
                          "name": {
                            "type": "string"
                          }
                        },
                        "required": ["name"],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "specialties": {
                      "items": {
                        "type": "string"
                      },
                      "type": "array"
                    },
                    "startsAt": {
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
                      "type": "string"
                    },
                    "state": {
                      "const": "ended",
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "slug",
                    "title",
                    "school",
                    "startsAt",
                    "specialties",
                    "speakers",
                    "state",
                    "recording"
                  ],
                  "type": "object"
                }
              ]
            },
            "type": "array"
          },
          "pagination": {
            "properties": {
              "hasMore": {
                "type": "boolean"
              },
              "nextCursor": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            },
            "required": ["nextCursor", "hasMore"],
            "type": "object"
          }
        },
        "required": ["data", "counts", "pagination"],
        "type": "object"
      },
...
      "UpcomingBroadcastListDto": {
        "items": {
          "properties": {
            "id": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            },
            "school": {
              "type": "string"
            },
            "slug": {
              "type": "string"
            },
            "speakers": {
              "items": {
                "properties": {
                  "name": {
                    "type": "string"
                  }
                },
                "required": ["name"],
                "type": "object"
              },
              "type": "array"
            },
            "specialties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "startsAt": {
              "format": "date-time",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
              "type": "string"
            },
            "state": {
              "enum": ["published", "live"],
              "type": "string"
            },
            "title": {
              "type": "string"
            }
          },
          "required": [
            "id",
            "slug",
            "title",
            "school",
            "startsAt",
            "specialties",
            "speakers",
            "state"
          ],
          "type": "object"
        },
        "type": "array"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "",
    "title": "DS Platform API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/directions/{id}": {
      "get": {
        "operationId": "DirectionsAdminController_detail",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["DirectionsAdmin"]
      },
      "patch": {
        "operationId": "DirectionsAdminController_update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["DirectionsAdmin"]
      }
    },
...
    "/v1/admin/event-directions/{id}": {
      "get": {
        "operationId": "EventDirectionsAdminController_detail",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["EventDirectionsAdmin"]
      }
    },
...
    "/v1/admin/events/{id}": {
      "get": {
        "operationId": "EventsAdminController_detail",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["EventsAdmin"]
      },
      "patch": {
        "operationId": "EventsAdminController_update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["EventsAdmin"]
      }
    },
...
    "/v1/public/directions/{idOrSlug}/events": {
      "get": {
        "operationId": "DirectionEventsPublicController_events",
        "parameters": [
          {
            "in": "path",
            "name": "idOrSlug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["DirectionEventsPublic"]
      }
    },
    "/v1/public/events": {
      "get": {
        "operationId": "EventsPublicController_list",
        "parameters": [
          {
            "in": "query",
            "name": "month",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "timeframe",
            "required": false,
            "schema": {
              "enum": ["upcoming", "past"],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/UpcomingBroadcastListDto"
                    },
                    {
                      "$ref": "#/components/schemas/MonthBroadcastListDto"
                    },
                    {
                      "$ref": "#/components/schemas/PublicEventListingPageDto"
                    }
                  ]
                }
              }
            },
            "description": ""
          }
        },
        "tags": ["EventsPublic"]
      }
    },
    "/v1/public/events/{idOrSlug}": {
      "get": {
        "operationId": "EventsPublicController_page",
        "parameters": [
          {
            "in": "path",
            "name": "idOrSlug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["EventsPublic"]
      }
    },
    "/v1/public/events/{idOrSlug}/directions": {
      "get": {
        "operationId": "EventDirectionsPublicController_directions",
        "parameters": [
          {
            "in": "path",
            "name": "idOrSlug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["EventDirectionsPublic"]
      }
    },
...
  },
  "servers": [],
...
}